Rename "dice" to "tcg"
Rename the source files for the TCG DICE extension in preparation to
adding non-TCG related stuff.
