+++ Makefile
@@ -3,20 +3,39 @@
 export GO111MODULE := on
 export SHELL := /bin/bash
 
-.PHONY: test
-test: ; @go test -v
+GOPKG := github.com/veraison/dice/tcg
 
-.PHONY: coverage
-coverage:
-	@go test -v -cover -race -coverprofile=coverage.out && \
-                go tool cover -html=coverage.out
-CLEANFILES += coverage.out
+GOLINT ?= golangci-lint
+
+ifeq ($(MAKECMDGOALS),lint)
+GOLINT_ARGS ?= run --timeout=3m -E dupl -E gocritic -E gosimple -E prealloc
+endif
 
 .PHONY: lint
-lint: ; @golangci-lint run
+lint: ; $(GOLINT) $(GOLINT_ARGS)
 
-.PHONY: clean
-clean: ; $(RM) -r $(CLEANFILES)
+ifeq ($(MAKECMDGOALS),test)
+GOTEST_ARGS ?= -v -race $(GOPKG)
+else
+  ifeq ($(MAKECMDGOALS),test-cover)
+  GOTEST_ARGS ?= -short -cover $(GOPKG)
+  endif
+endif
+
+COVER_THRESHOLD := $(shell grep '^name: cover' .github/workflows/ci-go-cover.yml | cut -c13-)
+
+.PHONY: test test-cover
+test test-cover: ; go test $(GOTEST_ARGS)
+
+presubmit:
+	@echo
+	@echo ">>> Check that the reported coverage figures are $(COVER_THRESHOLD)"
+	@echo
+	$(MAKE) test-cover
+	@echo
+	@echo ">>> Fix any lint error"
+	@echo
+	$(MAKE) lint
 
 .PHONY: licenses
 licenses: ; @./scripts/licenses.sh
@@ -24,10 +43,9 @@ licenses: ; @./scripts/licenses.sh
 .PHONY: help
 help:
 	@echo "Available targets:"
-	@echo
-	@echo "    test: run the package tests (default)"
-	@echo "coverage: run the package tests and show coverage profile"
-	@echo "    lint: run golangci-lint using configuration from .golangci.yml"
-	@echo "   clean: remove garbage"
-	@echo "licenses: check licenses of dependent packages"
-	@echo
+	@echo "  * test:       run unit tests for $(GOPKG)"
+	@echo "  * test-cover: run unit tests and measure coverage for $(GOPKG)"
+	@echo "  * licenses:   check licenses of dependent packages"
+	@echo "  * lint:       lint sources using default configuration"
+	@echo "  * presubmit:  check you are ready to push your local branch to remote"
+	@echo "  * help:       print this menu"
