+++ tcg/tcg_test.go
@@ -1,7 +1,7 @@
 // Copyright 2021 Contributors to the Veraison project.
 // SPDX-License-Identifier: Apache-2.0
 
-package dice
+package tcg
 
 import (
 	"io/ioutil"
